Use array expansion to add extra bazel args to the bazel invocation. (#17)
* Use array expansion to add extra bazel args to the bazel invocation.

This is necessary for cases when arguments are passed that involve quotation, such as "--ios_simulator_device=iPad Pro (12.9-inch)". If we store the extra args as a string, then when we expand it in the bazel invocation the string will be parsed as a string literal or as a whitespace-delimited array of arguments, which can cause the original argument quotations to be lost or changed. By expanding the arguments as an array, we honor the original quotations of the arguments.

* Remove bazel args variable.